Support CMAKE_INSTALL_LIBDIR and install into its python subdirectory
This makes the zeectl configuration honor the libdir configured in the
Zeek tree. We also no longer install into a fixed lib/zeekctl, and
instead install into a pyhon subdirectory in the inherited
libdir. This python directory becomes the common place for Python code
installed in the Zeek tree.

The old hardwired installation location is removed if still present.